Group models in dedicated namespace for DatabaseApi.
