+++ arangodb-net-standard.Test/Docs/UsageTest.cs
@@ -1,6 +1,7 @@
 ﻿using ArangoDBNetStandard;
 using ArangoDBNetStandard.CollectionApi.Models;
 using ArangoDBNetStandard.DatabaseApi;
+using ArangoDBNetStandard.DatabaseApi.Models;
 using ArangoDBNetStandard.Transport.Http;
 using System;
 using System.Collections.Generic;
